pyproject.toml
[project]
name = "dfs.recipes"
version = "0.1.0"
description = "Daily Fantasy Sports (DFS) Recipes"
authors = [{ name = "Anthony Nguyen", email = "anthony.nguyen636@gmail.com" }]
readme = "README.md"

requires-python = ">=3.12,<4.0.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "google-genai (>=1.19.0,<2.0.0)",
    "anthropic (>=0.54.0,<0.55.0)",
    "openai (>=1.86.0,<2.0.0)",
    "langchain[google-genai] (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "google-adk (>=1.2.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "langgraph-checkpoint-sqlite (>=2.0.10,<3.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "webauthn (>=2.5.2,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "slimit (>=0.8.1,<0.9.0)"
]

[project.urls]
repository = "https://github.com/an-vn/dfs.recipes"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
